+++ example_sync.py
@@ -14,15 +14,14 @@ def run():
         dpa.password = PASSWORD
 
         # Get LED settings of the device. The state might be LED_ON or LED_OFF.
-        # {'state': 'LED_ON'}
-        print(dpa.device.get_led_setting())
+        print("LED is on" if dpa.device.get_led_setting() else "LED if off")
 
         # Set LED settings of the device. Set enable to True to them turn on, to False to turn them off.
         # If the state was changed successfully, True is returned, otherwise False.
         print("success" if dpa.device.set_led_setting(enable=True) else "failed")
 
         # Factory reset the device. If the reset will happen shortly, True is returned, otherwise False.
-        # print("success" if dpa.device.factory_reset() else "failed")
+        print("success" if dpa.device.factory_reset() else "failed")
 
         # Restart the device. If the restart will happen shortly, True is returned, otherwise False.
         print("success" if dpa.device.restart() else "failed")
@@ -31,8 +30,9 @@ def run():
         print(dpa.device.uptime())
 
         # Check for new firmware versions
-        # {'result': 'UPDATE_NOT_AVAILABLE', 'new_firmware_version': ''}
-        print(dpa.device.check_firmware_available())
+        firmware = dpa.device.check_firmware_available()
+        print(firmware.result)  # "UPDATE_NOT_AVAILABLE"
+        print(firmware.new_firmware_version)  # ""
 
         # Start firmware update, if new version is available. Important: The response does not tell you anything about the
         # success of the update itself.
@@ -40,21 +40,20 @@ def run():
 
         # Get details of wifi stations connected to the device: MAC address, access point type (main or guest), wifi band and
         # connection rates.
-        # {'connected_stations':
-        #   [
-        #     {'mac_address': 'AA:BB:CC:DD:EE:FF',
-        #      'vap_type': 'WIFI_VAP_MAIN_AP',
-        #      'band': 'WIFI_BAND_5G',
-        #      'rx_rate': 87800,
-        #      'tx_rate': 87800}
-        #   ]
-        # }
-        print(dpa.device.get_wifi_connected_station())
+        connected_stations = dpa.device.get_wifi_connected_station()
+        print(connected_stations[0].mac_address)  # "AA:BB:CC:DD:EE:FF"
+        print(connected_stations[0].vap_type)  # "WIFI_VAP_MAIN_AP"
+        print(connected_stations[0].band)  # "WIFI_BAND_5G"
+        print(connected_stations[0].rx_rate)  # 87800
+        print(connected_stations[0].tx_rate)  # 87800
 
         # Get details about wifi guest access: SSID, Wifi key, state (enabled/disabled) and if time limited, the remaining
         # duration.
-        # {'ssid': 'devolo-guest-930', 'key': 'HMANPGBA', 'enabled': False, 'remaining_duration': 0}
-        print(dpa.device.get_wifi_guest_access())
+        guest_wifi = dpa.device.get_wifi_guest_access()
+        print(guest_wifi.ssid)  # "devolo-guest-930"
+        print(guest_wifi.key)  # "HMANPGBA"
+        print(guest_wifi.enabled)  # False
+        print(guest_wifi.remaining_duration)  # 0
 
         # Enable or disable the wifi guest access. Set enable to True to it turn on, to False to turn it off. Optionally
         # specify a duration in minutes. Changing SSID or the wifi key is currently not supported. If the state was changed
@@ -63,17 +62,13 @@ def run():
 
         # Get details about other access points in your neighborhood: MAC address, SSID, wifi band, used channel, signal
         # strength in DB and a value from 1 to 5, if you would want to map the signal strength to a signal bars.
-        # {'neighbor_aps':
-        #   [
-        #     {'mac_address': 'AA:BB:CC:DD:EE:FF',
-        #      'ssid': 'wifi',
-        #      'band': 'WIFI_BAND_2G',
-        #      'channel': 1,
-        #      'signal': -73,
-        #      'signal_bars': 1}
-        #   ]
-        # }
-        print(dpa.device.get_wifi_neighbor_access_points())
+        neighbor_aps = dpa.device.get_wifi_neighbor_access_points()
+        print(neighbor_aps[0].mac_address)  # "AA:BB:CC:DD:EE:FF"
+        print(neighbor_aps[0].ssid)  # "wifi"
+        print(neighbor_aps[0].band)  # "WIFI_BAND_2G"
+        print(neighbor_aps[0].channel)  # 1
+        print(neighbor_aps[0].signal)  # -73
+        print(neighbor_aps[0].signal_bars)  # 1
 
         # Start WPS push button configuration. If WPS was started successfully, True is returned, otherwise False.
         print("WPS started" if dpa.device.start_wps() else "WPS start failed")
@@ -81,45 +76,25 @@ def run():
         # Start WPS clone mode. If clone mode was started successfully, True is returned, otherwise False.
         print("WPS clone mode started" if dpa.device.start_wps_clone() else "WPS clone mode start failed")
 
-        # Get PLC network overview with enriched information like firmware version,
-        # {'network':
-        #   {'devices':
-        #       [
-        #         {'product_name': 'devolo dLAN pro 1200+ WiFi ac',
-        #          'product_id': 'MT2730',
-        #          'friendly_version': '2.8.0.01',
-        #          'full_version': 'MAC-QCA7500-2.8.0.30-01-20190705-CS',
-        #          'user_device_name': '',
-        #          'mac_address': 'AABBCCDDEEFF',
-        #          'topology': 'LOCAL',
-        #          'technology': 'HPAV_PANTHER',
-        #          'bridged_devices': [],
-        #          'user_network_name': '',
-        #          'ipv4_address': '',
-        #          'attached_to_router': False},
-        #         {'product_name': 'devolo dLAN 1200+',
-        #          'product_id': 'MT2639',
-        #          'friendly_version': '2.8.0.01-1',
-        #          'full_version': 'MAC-QCA7500-2.8.0.30-01-20190705-CS',
-        #          'user_device_name': '',
-        #          'mac_address': 'AABBCCDDEEFF',
-        #          'topology': 'REMOTE',
-        #          'technology': 'HPAV_PANTHER',
-        #          'bridged_devices': [],
-        #          'attached_to_router': True,
-        #          'user_network_name': '',
-        #          'ipv4_address': ''}
-        #       ],
-        #    'data_rates':
-        #      [
-        #        {'mac_address_from': 'AABBCCDDEEFF',
-        #         'mac_address_to': 'AABBCCDDEEFF',
-        #         'tx_rate': 129.9375,
-        #         'rx_rate': 124.6875}
-        #      ]
-        #   }
-        # }
-        print(dpa.plcnet.get_network_overview())
+        # Get PLC network overview with enriched information like firmware version.
+        network = dpa.plcnet.get_network_overview()
+        print(network.devices[0].product_name)  # "devolo Magic 2 WiFi next"
+        print(network.devices[0].product_id)  # "MT3056"
+        print(network.devices[0].friendly_version)  # "7.12.5.124"
+        print(network.devices[0].full_version)  # "magic-2-wifi-next 7.12.5.124_2022-08-29"
+        print(network.devices[0].full_version)  # "magic-2-wifi-next 7.12.5.124_2022-08-29"
+        print(network.devices[0].user_device_name)  # "Living Room"
+        print(network.devices[0].mac_address)  # "AABBCCDDEEFF"
+        print(network.devices[0].topology)  # devolo_plc_api.plcnet_api.Device.LOCAL
+        print(network.devices[0].technology)  # devolo_plc_api.plcnet_api.Device.GHN_SPIRIT
+        print(network.devices[0].bridged_devices)  # []
+        print(network.devices[0].attached_to_router)  # True
+        print(network.devices[0].user_network_name)  # ""
+        print(network.devices[0].ipv4_address)  # ""
+        print(network.data_rates[0].mac_address_from)  # "AABBCCDDEEFF"
+        print(network.data_rates[0].mac_address_to)  # "AABBCCDDEEFF"
+        print(network.data_rates[0].tx_rate)  # 129.9375
+        print(network.data_rates[0].rx_rate)  # 124.6875
 
         # Identify the device by making the PLC LED blink. This call returns directly with True, if identifying was started
         # successfully, otherwise False. However, the LED stays blinking for two minutes.