+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Specify a duration for the guest wifi
 - Start WPS clone mode
+- Factory reset device
 
 #### PLC API