+++ docs/CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v1.0.0] - 2023/01/05
 
 ### Changed
 
-- **BREAKING**: The results are now dataclass-like objects
+- **BREAKING**: The results are now dataclass-like objects. Please have a look at our [examples](https://github.com/2Fake/devolo_plc_api/blob/dee4617da680685a35ac48051d1aecd0456b7764/example_async.py) to see, how migration works.
 
 ## [v0.9.0] - 2022/12/20